Did some styling and refactoring.
Deleted the non necessary stuff in the css and the things that were not used
in the ts files. Refactored some methods and made it so that the site reloads
when a user deletes a post so that the new view shows that the post is gone.
Deleted the components that were not used anymore.